refactor: 랜딩 페이지 섹션 컴포넌트 분리
